+++ verl/verl/utils/megatron/tensor_parallel.py
@@ -31,7 +31,7 @@ def update_kwargs_with_config(dictionary: Dict, config: ModelParallelConfig):
 
 def get_default_kwargs_for_model_parallel_config():
     model_parallel_config_kwargs = {
-        'params_dtype': torch.float32,
+        'params_dtype': torch.bfloat16,
         'use_cpu_initialization': False,
         'perform_initialization': True,
         'gradient_accumulation_fusion': False,
